move VespaResult from query module to application module
